+++ packages/relay/tests/lib/eth/eth_gasPrice.spec.ts
@@ -36,7 +36,7 @@ import HAPIService from '../../../src/lib/services/hapiService/hapiService';
 import HbarLimit from '../../../src/lib/hbarlimiter';
 import { numberTo0x } from '../../../dist/formatters';
 import { CacheService } from '../../../src/lib/services/cacheService/cacheService';
-import { DEFAULT_NETWORK_FEES, ETH_FEE_HISTORY_VALUE } from './config';
+import { DEFAULT_NETWORK_FEES, ETH_FEE_HISTORY_VALUE } from './eth-config';
 import { predefined } from '../../../src/lib/errors/JsonRpcError';
 import RelayAssertions from '../../assertions';
