+++ packages/relay/tests/lib/eth/eth_getBlockByHash.spec.ts
@@ -60,7 +60,7 @@ import {
   ETH_FEE_HISTORY_VALUE,
   LINKS_NEXT_RES,
   NO_SUCH_BLOCK_EXISTS_RES,
-} from './config';
+} from './eth-config';
 
 dotenv.config({ path: path.resolve(__dirname, '../test.env') });
 chai.use(chaiAsPromised);
