+++ packages/relay/tests/lib/eth/eth_getBlockTransactionCountByNumber.spec.ts
@@ -45,7 +45,7 @@ import {
   DEFAULT_NETWORK_FEES,
   ETH_FEE_HISTORY_VALUE,
   NO_SUCH_BLOCK_EXISTS_RES,
-} from './config';
+} from './eth-config';
 
 dotenv.config({ path: path.resolve(__dirname, '../test.env') });
 chai.use(chaiAsPromised);
